Refactor code generation and add chat command
- Move saveGeneratedFiles function to string_utils.ts
- Extract generateCommitMessagePrompt to git_utils.ts
- Add new chat command to CLI
- Update imports and function calls accordingly
