+++ src/index.ts
@@ -8,6 +8,7 @@ import init from './commands/init'
 import proxy from './commands/proxy'
 import tokens from './commands/tokens'
 import models from './commands/models'
+import chat from './commands/chat'
 import shell from './commands/generate/shell'
 import code from './commands/generate/code'
 import commit from './commands/generate/commit'
@@ -44,6 +45,14 @@ program
   .option('-s, --silent', 'Run the command without asking for confirmation of submitting prompts')
   .option('-m, --model <model>', 'Select the model to use (e.g., openai, claude, ollama, ollama:llama3, claude:claude-3-5-sonnet-20240620)')
   .action(init)
+
+program
+  .command('chat')
+  .option('-f, --files <filenames...>', 'Specify filenames of files with context to use for generating code')
+  .option('-t, --temperature <temperature>', 'Temperature for the model (default 0.2)')
+  .option('-m, --model <model>', 'Select the model to use (e.g., openai, claude, ollama, ollama:llama3, claude:claude-3-5:sonnet-20240620)')
+  .description('Start an interactive chat with cloving')
+  .action(chat)
 
 program
   .command('proxy')
