+++ src/utils/git_utils.ts
@@ -62,3 +62,23 @@ export const getCurrentBranchName = (): string => {
     process.exit(1)
   }
 }
+
+export const generateCommitMessagePrompt = (diff: string): string => {
+  return `Generate a concise and meaningful commit message based on a diff.
+
+Do not add any commentary or context to the message other than the commit message itself.
+
+An example of the output for this should look like the following:
+
+\`\`\`plaintext
+Update dependencies and package versions
+
+- Upgrade Ruby gems including aws-sdk, honeybadger, irb, and rubocop
+- Update Node.js packages including esbuild, tinymce, and trix
+- Bump TypeScript and ESLint related packages to latest versions
+\`\`\`
+
+Here is the diff to help you write the commit message:
+
+${diff}`
+}
