+++ src/utils/string_utils.ts
@@ -1,3 +1,6 @@
+import fs from 'fs'
+import path from 'path'
+
 // Function to estimate token count
 export const estimateTokens = async (text: string): Promise<number> => {
   const charCount = text.length
@@ -123,4 +126,17 @@ export const extractFilesAndContent = (rawCodeCommand: string | undefined): [str
   }
 
   return [files, fileContents]
+}
+
+export const saveGeneratedFiles = async (files: string[], fileContents: Record<string, string>): Promise<void> => {
+  for (const file of files) {
+    if (fileContents[file]) {
+      const filePath = path.resolve(file)
+      await fs.promises.mkdir(path.dirname(filePath), { recursive: true })
+      await fs.promises.writeFile(filePath, fileContents[file])
+      console.log(`${file} has been saved.`)
+    } else {
+      console.log(`File content not found for ${file}.`)
+    }
+  }
 }
